refactor(desktop): move device lock into view model
Keep the LWD executor view free of external system hooks.
